requirements.txt
opencv-python
numpy
Pillow
face-recognition
face_recognition_models
setuptools
mariadb
boto3